==> notebooks/01_data_ingestion.ipynb <==
from datetime import datetime, timedelta

import polars as pl
import requests

#### Fetch OPW Version

dt_end = datetime.today() - timedelta(days=1)

url = (
    "https://api.data.gov.hk/v1/historical-archive/list-file-versions?"
    "url=https://online-price-watch.consumer.org.hk/opw/opendata/pricewatch.json&"
    "start={}&end={}&"
)

versions = []

for delta in range(0, 30, 10):
    end = dt_end - timedelta(days=delta)
    start = end - timedelta(days=9)
    start_end = start.strftime("%Y%m%d"), end.strftime("%Y%m%d")

    response = requests.get(url.format(*start_end))
    response.raise_for_status()

    data = response.json()
    versions += data.get("timestamps", [])

date_version = dict(
    pl.DataFrame({"version": versions})
    .with_columns(
        pl.col("version").str.slice(0, 8)
            .str.to_date("%Y%m%d")
            .alias("date")
    )
    .with_columns(
        (pl.col("date") - pl.duration(days=1))
            .dt.strftime("%Y%m%d")
    )
    .select("date", "version")
    .iter_rows()
)

#### Download OPW Data

url = (
    "https://api.data.gov.hk/v1/historical-archive/get-file?"
    "url=https://online-price-watch.consumer.org.hk/opw/opendata/pricewatch.json&"
    "time={}&"
)

prices, items = [], []

for date, version in date_version.items():
    response = requests.get(url.format(version))
    response.raise_for_status()
    data = response.json()

    for item in data:
        item["code"] = str(item["code"]).upper()
        code = item["code"]

        price = item.pop("prices", [])
        offer = item.pop("offers", [])

        smkt_price = {p["supermarketCode"]: p for p in price}
        smkt_offer = {o["supermarketCode"]: o for o in offer}

        price = [
            {
                "code": code, "date": date,
                **smkt_price.get(smkt, {}), **smkt_offer.get(smkt, {}),
            }
            for smkt in set(smkt_price) | set(smkt_offer)
        ]

        prices += price
        items.append(item)

df_item = pl.json_normalize(items)
df_price = pl.from_records(prices)

df_item.write_parquet("../data/raw/items.parquet")
df_price.write_parquet("../data/raw/prices.parquet")
