==> notebooks/03_database_functions.ipynb <==
import json
import os

import polars as pl
from dotenv import load_dotenv
from supabase import create_client
from supabase.client import ClientOptions


load_dotenv("../config/.env")

df_item = pl.read_parquet("../data/cleansed/items.parquet")
df_price = pl.read_parquet("../data/cleansed/prices.parquet")

url = os.getenv("SUPABASE_URL")
key = os.getenv("SUPABASE_KEY")
schema = os.getenv("SUPABASE_SCHEMA")
client = create_client(url, key, ClientOptions(schema=schema))

#### Insert Data

data = df_price.write_json()
data = json.loads(data)

client.table("prices").insert(data).execute()

#### Delete Data

client.table("prices").delete().in_("effective_date", ["20240101"]).execute()

#### Call Functions

client.rpc("check_connection").execute()

client.rpc("remove_user", {"usr_id": 1}).execute()
