==> notebooks/04_bot_message.ipynb <==
import json
import os

import requests
from dotenv import load_dotenv


load_dotenv("../config/.env")

chat_id = os.getenv("TELEGRAM_TESTER_ID")
token = os.getenv("TELEGRAM_TOKEN")
url_base = f"https://api.telegram.org/bot{token}"
url_msg = f"{url_base}/sendMessage"
url_img = f"{url_base}/sendPhoto"
url_edit = f"{url_base}/editMessageText"

#### Send Message

data = {
    "chat_id": chat_id,
    "text": "I'm here to save your money! 🐶",
}

requests.get(url_msg, data).json()

#### Send Image

img = open("test.png", "rb")

data = {
    "chat_id": chat_id,
}

files = {
    "photo": img,
}

requests.post(url_img, data, files=files).json()

#### Send Inline Keyboard

keyboard = json.dumps({
    "inline_keyboard": [
        [
            {"text": "watchdog", "callback_data": "edit_watchdog"},
            {"text": "subscription", "callback_data": "edit_sub"},
        ],
        [
            {"text": "setting", "callback_data": "edit_setting"},
            {"text": "nothing", "callback_data": "nothing"},
        ],
    ]
})

data = {
    "chat_id": chat_id,
    "text": "helping you",
    "parse_mode": "HTML",
    "reply_markup": keyboard,
}

requests.post(url_msg, data).json()

#### Edit Inline Keyboard

keyboard = json.dumps({
    "inline_keyboard": [[]]
})

data = {
    "chat_id": chat_id,
    "message_id": 288,
    "text": "helping you",
    "parse_mode": "HTML",
    "reply_markup": keyboard,
}

requests.post(url_edit, data).json()

#### Remove Keyboard

data = {
    "chat_id": chat_id,
    "text": "Here is a message with the keyboard removed.",
    "reply_markup": {
        "remove_keyboard": True,
    },
}

response = requests.post(url_msg, json=data)
